Feature/fix mac playbook: Add CI environment specific playbook for Mac

+++ scripts/distributions/Mac/uninstall-mono-mdk.sh
@@ -10,6 +10,8 @@ sudo rm -r /Library/Receipts/MonoFramework-*
 for dir in /usr/bin /usr/share/man/man1 /usr/share/man/man3 /usr/share/man/man5; do
     (cd ${dir};
     for i in `ls -al | grep /Library/Frameworks/Mono.framework/ | awk '{print $9}'`; do
-        sudo rm ${i}
+        if [ -n "${i}" ]; then
+            sudo rm ${i}
+        fi
     done);
 done
